OK, shipping the raw jars is too damned big. So I took another crack at convincing GWT and Proguard to play nicely together. In spite of hours more of fucking around trying to figure out what of the 1.2MB gwt-servlet.jar is actually necessary, I just ended up including the whole damned thing.
Why do I always think that it would be great to the moral equivalent of hooking
a diesel engine up to a toaster? They obviously did not design the diesel
engine to be used indoors, nor to have its output harnessed by a home
appliance.

Anyway, now the client is only 5ish MB instead of like 20.