Two bits:
1. Fixed up Squeryl so that I can do everything I want with foreign key
constraints, so now when I want to delete a project, I just delete the project
and everything else is magically deleted thanks to cascading keys.

2. Totally revamped the (fqName -> id) mapping to use the Neo4j graph database.
This is about a jillion times faster than the naive massive flattened mapping
we used to use, and a jillion times faster than an intermediate approach (which
I never committed to version control) that used an (id, name, parentId) SQL
table and wired everything up manually with a ton of queries. It may be that a
graph database is more appropriate for everything, but we'll fry that fish
later. For now I'm just happy that imports are a good bit faster (which will
hopefully take the edge off the fact that I think I'm going to have to force a
database wipe with this release).